Timezone is not required anymore.
Originally the timezone was added to make sure we get all games for a certain date.
The script was run once a day and retrieved all games for the previous day.
The script can now be run any amount of times per day as it processes games
in a 2 day bracket (today and yesterday), moving forward 1 day at a time.
This overlap ensures that we get all games.